Update MQ2AutoClaim
Used the wrong / for the dates and the parser was looking for / but I was using \, which messed up parsing of the dates. 
Fixed the issue, and made it so that it parses either slash so it would parse existing INI's with the error in them as opposed to asking everyone to delete their INI's.